# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quickstart in Couchbase with PHP and Laravel",
...
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/v1/airlines/list": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get list of airlines",
                "description": "Returns list of airlines",
                "operationId": "getAirlinesList",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "United States"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airline"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airlines/{id}": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get airline information",
                "description": "Returns airline data",
                "operationId": "getAirlineById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Airline"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Update an existing airline or create a new one if it does not exist",
                "description": "Update an existing airline or create a new one if it does not exist",
                "operationId": "updateAirline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Airline object that needs to be updated or created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Airline"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Resource created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Create a new airline",
                "description": "Create a new airline",
                "operationId": "createAirline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Airline object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Airline"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Delete an airline",
                "description": "Delete an airline",
                "operationId": "deleteAirline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airlines/to-airport": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get airlines flying to a destination airport",
                "description": "Returns list of airlines flying to a specific airport",
                "operationId": "getAirlinesToAirport",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Airline": {
                "title": "Airline",
                "required": [
                    "callsign",
                    "country",
                    "iata",
                    "icao",
                    "name"
                ],
                "properties": {
                    "callsign": {
                        "description": "Callsign of the airline",
                        "type": "string"
                    },
                    "country": {
                        "description": "Country of the airline",
                        "type": "string"
                    },
                    "iata": {
                        "description": "IATA code of the airline",
                        "type": "string"
                    },
                    "icao": {
                        "description": "ICAO code of the airline",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of the airline",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}